abstract components refactored
