+++ projects/opal-frontend-common/components/abstract/abstract-form-alias-base/interfaces/abstract-form-alias-base-alias-controls.interface.ts
@@ -1,4 +1,4 @@
-import { IAbstractFormArrayControl } from '../../interfaces/abstract-form-array-control.interface';
+import { IAbstractFormArrayControl } from '@hmcts/opal-frontend-common/components/abstract/interfaces';
 
 export interface IAbstractFormAliasBaseAliasControls {
   [key: string]: IAbstractFormArrayControl;
