+++ projects/opal-frontend-common/components/abstract/abstract-form-alias-base/public-api.ts
@@ -1,0 +1,1 @@
+export * from './abstract-form-alias-base.component';
