+++ projects/opal-frontend-common/components/abstract/abstract-form-array-base/interfaces/abstract-form-array-base-form-array-controls.interface.ts
@@ -1,4 +1,4 @@
-import { IAbstractFormArrayControl } from '@components/abstract/interfaces/abstract-form-array-control.interface';
+import { IAbstractFormArrayControl } from '@hmcts/opal-frontend-common/components/abstract/interfaces';
 
 export interface IAbstractFormArrayBaseFormArrayControls {
   [key: string]: IAbstractFormArrayControl;
