+++ projects/opal-frontend-common/components/abstract/abstract-form-array-removal-base/public-api.ts
@@ -1,0 +1,1 @@
+export * from './abstract-form-array-removal-base.component';
