+++ projects/opal-frontend-common/components/abstract/abstract-form-parent-base/package.json
@@ -1,0 +1,6 @@
+{
+    "sideEffects": false,
+    "module": "../../../../dist/opal-frontend-common/fesm2022/hmcts-opal-frontend-common-components-abstract-abstract-form-parent-base.mjs",
+    "typings": "../../../../dist/opal-frontend-common/components/abstract/abstract-form-parent-base/index.d.ts"
+  }
+
