+++ projects/opal-frontend-common/components/abstract/abstract-sortable-table-pagination/public-api.ts
@@ -1,0 +1,1 @@
+export * from './abstract-sortable-table-pagination.component';
