--- projects/opal-frontend-common/components/abstract/package.json
@@ -1,6 +1,0 @@
-{
-    "sideEffects": false,
-    "module": "../../../../dist/opal-frontend-common/fesm2022/hmcts-opal-frontend-common-components-abstract.mjs",
-    "typings": "../../../../dist/opal-frontend-common/components/abstract/index.d.ts"
-  }
-
