--- projects/opal-frontend-common/components/abstract/public-api.ts
@@ -1,24 +1,0 @@
-// ABSTRACT CLASSES
-export * from './abstract-form-alias-base/abstract-form-alias-base';
-export * from './abstract-form-array-base/abstract-form-array-base';
-export * from './abstract-form-array-parent-base/abstract-form-array-parent-base.component';
-export * from './abstract-form-array-removal-base/abstract-form-array-removal-base.component';
-export * from './abstract-form-base/abstract-form-base.component';
-export * from './abstract-form-parent-base/abstract-form-parent-base.component';
-export * from './abstract-sortable-table/abstract-sortable-table.component';
-export * from './abstract-sortable-table-pagination/abstract-sortable-table-pagination.component';
-
-// INTERFACES
-export * from './abstract-form-base/interfaces/abstract-form-base-field-error.interface';
-export * from './abstract-form-base/interfaces/abstract-form-base-field-errors.interface';
-export * from './interfaces/abstract-form-base-form-error-summary-message.interface';
-export * from './abstract-form-base/interfaces/abstract-form-base-form-error.interface';
-export * from './abstract-form-base/interfaces/abstract-form-base-form.interface';
-export * from './abstract-form-base/interfaces/abstract-form-base-high-priority-form-error.interface';
-export * from './abstract-form-base/interfaces/abstract-form-base-status.interface';
-export * from './interfaces/abstract-form-array-control-validation.interface';
-export * from './interfaces/abstract-form-array-control.interface';
-export * from './interfaces/abstract-form-array-controls.interface';
-export * from './interfaces/abstract-form-base-form-error-summary-message.interface';
-export * from './interfaces/abstract-form-control-error-message.interface';
-export * from './abstract-sortable-table/interfaces/abstract-sortable-table-interfaces';
