+++ projects/opal-frontend-common/components/govuk/govuk-error-summary/govuk-error-summary.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { IAbstractFormBaseFormErrorSummaryMessage } from '@hmcts/opal-frontend-common/components/abstract';
+import { IAbstractFormBaseFormErrorSummaryMessage } from '@hmcts/opal-frontend-common/components/abstract/interfaces';
 
 @Component({
   selector: 'opal-lib-govuk-error-summary',
